Use Android's timely frame posting
Android's SurfaceFlinger can do the timing of the frames for us
Compute the microsecond-precise frame timestamp, and give that us timestamp to Android.

Before that, we were doing millisecond-precise computations, only using
sleeps (so system wouldn't fix if missed some steps, and far from
real-time so it was very approximative)

SurfaceFlinger has a high enough priority to compute the frames at each
vsync, so it'll be smoother than us.