Revamp passthrough mode 1 on modern Android
The idea of mode 1 was, and remain, that avos/ffmpeg does the
spdif/IEC61937 encapsulation, while with mode 2, encapsulation is done
by the system.
This supposes that the system knows how to encapsulate, and knows how to
do it properly.

Behavior on old Android version remain the same, open a direct PCM
audiotrack (or even not direct)
But on recent Android versions, we can instead open an IEC61937 audio
track, which is a better guarantee it'll actually passthrough (still not
perfect).

Since DTS-HD can be downgraded to DTS, we need to know the maximum
IEC61937 baudrate, which Android doesn't provide.
We use some proxy information from supported audio codecs to determine
which is supported.
I'm expecting this is missing some steps, like supporting the
intermediate baudrate 192kHz, but I'll expect user reports.
Either way, we're expecting the vast majority of our users to use mode
2, which isn't modified by this change