Adjust frame duration to improve playback fluidity when using android synchronization
When playing a 60Hz stream 1ms timescale on a 60Hz screen, frame length 17ms 17ms 16ms are obtained.
This results roughly in one chance in 20 to render it as 2 frames 1 frame 0 frame, rather than 1 1 1.
Check that if frame duration matches a repeating pattern to replace the frame duration with the average of the frame duration during one pattern length.

Credits to phh for this contribution and analysis.